UHF-89: Abstract translatable fields to make it more re-usable
